refactor: 패키지 분리 및 안 쓰는 코드 제거
