+++ README.md
@@ -5,6 +5,7 @@
 - 카드
     - Denomination(끗수): 2,3,4,5,6,7,8,9,10,J,Q,K,A
     - Suit(슈트): 스페이드, 하트, 다이아몬드, 클럽
+    - Hand(핸드): 카드 패 뭉치
 
 - 액션
     - Hit(힛): 카드를 더 받겠다는 의미
