--- src/main/java/blackjack/model/CardGenerator.java
@@ -1,5 +1,0 @@
-package blackjack.model;
-
-public interface CardGenerator {
-    Card pick();
-}
