--- src/main/java/blackjack/model/NumberGenerator.java
@@ -1,6 +1,0 @@
-package blackjack.model;
-
-@FunctionalInterface
-public interface NumberGenerator {
-    int pick();
-}
