--- src/main/java/blackjack/model/RandomNumberGenerator.java
@@ -1,10 +1,0 @@
-package blackjack.model;
-
-import java.util.Random;
-
-public class RandomNumberGenerator implements NumberGenerator {
-    @Override
-    public int pick() {
-        return new Random().nextInt(1000);
-    }
-}
