+++ src/main/java/blackjack/model/card/Card.java
@@ -1,0 +1,19 @@
+package blackjack.model.card;
+
+public class Card {
+    private final Suit suit;
+    private final Denomination denomination;
+
+    public Card(final Suit suit, final Denomination denomination) {
+        this.suit = suit;
+        this.denomination = denomination;
+    }
+
+    public Suit getSuit() {
+        return suit;
+    }
+
+    public Denomination getDenomination() {
+        return denomination;
+    }
+}
