+++ src/main/java/blackjack/model/card/Suit.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.card;
 
 public enum Suit {
     HEART,
