+++ src/main/java/blackjack/model/cardgenerator/CardGenerator.java
@@ -1,0 +1,7 @@
+package blackjack.model.cardgenerator;
+
+import blackjack.model.card.Card;
+
+public interface CardGenerator {
+    Card pick();
+}
