+++ src/test/java/blackjack/model/card/CardTest.java
@@ -1,9 +1,9 @@
-package blackjack.model;
-
-import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+package blackjack.model.card;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class CardTest {
     @Test
