+++ src/test/java/blackjack/model/card/HandTest.java
@@ -1,10 +1,11 @@
-package blackjack.model;
+package blackjack.model.card;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class HandTest {
     @Test
