+++ src/test/java/blackjack/model/dealer/DealerTest.java
@@ -1,10 +1,15 @@
-package blackjack.model;
+package blackjack.model.dealer;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.List;
+import blackjack.model.card.Card;
+import blackjack.model.card.Denomination;
+import blackjack.model.card.Suit;
+import blackjack.model.cardgenerator.SequentialCardGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class DealerTest {
     @Test
