+++ src/test/java/blackjack/model/player/PlayerTest.java
@@ -1,11 +1,16 @@
-package blackjack.model;
+package blackjack.model.player;
+
+import blackjack.model.card.Card;
+import blackjack.model.card.Denomination;
+import blackjack.model.card.Suit;
+import blackjack.model.cardgenerator.SequentialCardGenerator;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import java.util.List;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 
 public class PlayerTest {
     @Test
